Add debug utils to trigger clean shutdown/reboot
They wrap the calls in fbpad, which halt.sh specifically looks for and
doesn't kill. This allows the full shutdown script output (all the way
up to the final `poweroff -f` or `reboot -f`) to be seen and recorded.
(Probably we don't need this long term, but it's handy for testing.)