Remove spurious cat breaking condition in power.sh
The current code intends to disable soft sleep/wake for certain
foreground_process values, specifically muxcharge. The condition is
wrong for the sleep codepath, but correct for the wake codepath, so
currently, when the lid is SP lid is closed when charging, we SIGSTOP
muxcharge, but never SIGCONT it. The only way to get the device out of
this state is a hard reset.