Only put SP in sleep-closed state if it's asleep
I think originally, the SP power.sh script put the device to sleep every
time the lid was closed, so this check wasn't needed. But now the script
sometimes _doesn't_ put the device to sleep, so we need to check or else
we can get into the sleep-closed state when the device is awake... which
means we'll stop processing hotkeys when we should, and other weirdness.