Handle disabled sleep, suspend-to-RAM on lid close
This partially duplicates logic that's normally handled by input.sh for
power button presses, since the SP has special logic in power.sh to poll
the hallkey file for the lid switch.

This isn't great, but should hopefully tide us over and resolve the
remaining inconsistent sleep behavior on the SP till we can do a deeper
rework of the hallkey handling.