Fix known bugs in RG35XX-SP lid sleep (#310)
* Remove spurious cat breaking condition in power.sh

The current code intends to disable soft sleep/wake for certain
foreground_process values, specifically muxcharge. The condition is
wrong for the sleep codepath, but correct for the wake codepath, so
currently, when the lid is SP lid is closed when charging, we SIGSTOP
muxcharge, but never SIGCONT it. The only way to get the device out of
this state is a hard reset.

* Only put SP in sleep-closed state if it's asleep

I think originally, the SP power.sh script put the device to sleep every
time the lid was closed, so this check wasn't needed. But now the script
sometimes _doesn't_ put the device to sleep, so we need to check or else
we can get into the sleep-closed state when the device is awake... which
means we'll stop processing hotkeys when we should, and other weirdness.

* Handle disabled sleep, suspend-to-RAM on lid close

This partially duplicates logic that's normally handled by input.sh for
power button presses, since the SP has special logic in power.sh to poll
the hallkey file for the lid switch.

This isn't great, but should hopefully tide us over and resolve the
remaining inconsistent sleep behavior on the SP till we can do a deeper
rework of the hallkey handling.