Move state slot hotkeys into retroarch.device.cfg
They don't currently vary across devices, but that might change with
future device support, and in any case it's more consistent since all
the other hotkeys are in the device config now.

This change is backwards compatible; it's fine for someone to have the
same option set in both retroarch.cfg (which takes precedence) and the
included retroarch.device.cfg.