Fix backup folder path for RA config restore
Third time's the charm? Done on mobile and untested, but I think this is the ticket. (Directories are ending up in the image now, but I messed up the path. This new location matches the one that's actually in the script.)

P.S. Done quickly on a phone as a passenger in the car, so don't flame me too much for sending a PR while on vaca, lol.