Pass hotkey.json to muhotkey to parse directly
I tweaked the JSON format to group the hotkeys in an object rather than
an array, since the accompanying change to muhotkey causes it to sort
the hotkeys correctly (from longest to shortest) automatically.

I also removed the copying of hotkey.json from
/opt/muos/device/current/control to /opt/muos/config, since this will
make incremental updates that add hotkeys a bit simpler. (We can revisit
this in the future if we ever do want to add a GUI for hotkey editing.)