Show progress bars for archive installs
Adds pv (thanks @xongledongle for introducing me to this!!) and uses it
to show some nice progress bars for both the unzip and the file move.

Incidentally, this seems to make the archive install a fair bit faster.
I think we were actually getting bogged down printing lines to fbpad. :O