Change close_game.sh from sourced to standalone
We used to source close_game.sh and rely on it setting variables for the
calling script, but no longer do that, so it's cleaner to have it just
run as a standalone script.

Additionally, renamed it to quit.sh to reflect the more general purpose,
and fixed a bug where sometimes fbpad closes too early with verbose
messages enabled (e.g., in the new update flow).