Save default values of new vars after patch update
When a new variable is added in a patch archive,
/opt/muos/config/config.ini or /opt/muos/device/current/config.ini is
replaced with the default config file (including the new var's default).

Then, when we reboot after the update, the var save script replaces that
config.ini in its entirety with the values from /run/muos. This is good,
since it means we don't override the user's config customizations on
update... but since the newly added variable doesn't have a value in
/run/muos at yet, we ignore its default and set it to the empty string.

The way around this is to update the variable save routines to preserve
the existing value from config.ini if the relevant var doesn't exist in
/run/muos yet. (This does require writing the config to a new file and
moving it over config.ini instead of clearing the file and rebuilding
in place, but that actually might be better anyway, since it makes the
config.ini update atomic.)