Added MinUI Directory Calls to assign.json (#471)
* Added MinUI Directory Calls to assign.json

Added the code needed for MuOS to recognize the ROM directory names used in MinUI.  These directories include:

- Game Boy (GB)
- Game Boy Advance (GBA)
- Game Boy Color (GBC)
- Nintendo Entertainment System (FC)
- Sega Genesis (MD)
- Sony PlayStation (PS)
- Super Nintendo Entertainment System (SFC)
- Game Boy Advance (MGBA)
- Neo Geo Pocket Color (NGPC)
- Pico-8 (P8)
- Pokémon mini (PKM)
- Sega Game Gear (GG)
- Sega Master System (SMS)
- Super Game Boy (SGB)
- Super Nintendo Entertainment System (SUPA)
- TurboGrafx-16 (PCE)
- Virtual Boy (VB)

* Removed Spaces, Upper Case, Dashes & Accent Mark

As discussed on #471, I removed the spaces, upper case, dashes and accent mark.

* Added Dashes, Accent Mark

Added lines for Pico-8, Pokémon Mini and TurboGrafx-16 with dashes and accent mark that are consistent with MinUI naming.