Add support for folder-style PSP games
PSP games in a folder with an EBOOT.PBP file aren’t as well supported today as
games as a disk image, as they have to be placed in the PPSSPP memstick folder,
then you have to run PPSSPP and launch it from within the emulator. While this
works, it isn’t well integrated like other content is in muOS. While most
release games can be trivially converted to a disk image format, it doesn’t seem
like this is a good option for homebrew games like Minecraft PSP or others, so
those have to be run directly from PPSSPP at this time.

Instead, we can make use of a hidden folder and a placeholder file the way it’s
done for some other systems like ScummVM. If a `<game name>.psp` file is used,
it will check for a hidden game folder of the same name, a folder in .hidden/, or
a folder in the PPSSPP memstick folder, and run the emulator with that instead.
This lets the user play homebrew PSP games directly from the content explorer
rather than launching them directly from PPSSPP.