Reducing Boot Times due to Network Module (#635)
* Implemented Network Module Retry on Reload

Shortened the MAX_WAIT and added MAX_RETRY (which is set to the same
value as MAX_WAIT as is)

Added a retry loop to the reload-network function

* Implemented new configs for module loading

Added new config variables for MAX_WAIT (wait_timer) and MAX_RETRY
(compat_retry) to make the network module loading behaviour
configurable.

This configurability will enable users to tweak the behaviour for their
device and their preference. Defaults to 5 retries and a 5 second timer
as a good enough default most devices should be okay with.